[ Fix ]:  1. Delete duplicate files 2. Integrate codes
